Updating documentation to remove using fail* API in examples
The fail* API are deprecated, but were the primary API used in the
examples. This commit changes this, by using API that is recommended
for use with new projects. The fail* API are mentioned with the
remaining convenience API, but are introduced as being deprecated.

Additionally, updated what version and on what system the examples
were built against.

git-svn-id: http://svn.code.sf.net/p/check/code/trunk@678 64e312b2-a51f-0410-8e61-82d0ca0eb02a