Remove warnings from adding Wstrict-aliasing
The functions that used ppack were attempting to cast whatever was
being passed to ppack as a CheckMsg*. However, the things being
cast where in some cases not even structures the same size as
a CheckMsg. Really what the functions were attempting was to
pass a buffer and a type, which would be deciphered.

This expected datatype of ppack, and the eventual functions which
do the packing, now take void*. The expected data structures are
cast from what was passed.

git-svn-id: http://svn.code.sf.net/p/check/code/branches/warnings@705 64e312b2-a51f-0410-8e61-82d0ca0eb02a