Do no mix const and non-const data in check_check_pack unit tests
The msg containers in the unit test previously held a combination of
const and allocated data. Const strings cannot be assigned to these
containers if the const rules are to be strictly followed. To prevent
const compiler warnings, all const strings are now allocated strings,
which are free'd when they are no longer useful.

git-svn-id: http://svn.code.sf.net/p/check/code/branches/warnings@727 64e312b2-a51f-0410-8e61-82d0ca0eb02a